requirements.txt
fastapi
uvicorn
redis
google-cloud-storage
google-generativeai
python-dotenv
python-multipart
black
isort
